CSS-in-JS: Item checkbox
