+++ src/index.tsx
@@ -11,7 +11,7 @@ export { default as DropdownButton } from 'dropdownButton'
 export { default as EmptyState } from 'emptyState'
 export { default as HamburgerButton } from 'hamburgerButton'
 export { default as ItemAction } from 'itemAction'
-export { default as ItemCheckbox } from 'itemCheckbox'
+export { default as ItemCheckbox, ItemCheckboxStatus } from 'itemCheckbox'
 export { default as ItemChoice } from 'itemChoice'
 export { default as ItemData } from 'itemData'
 export { default as ItemInfo } from 'itemInfo'
